# preallocate buffer space